backend/requirements/local.txt
-r base.txt
django-cors-headers==3.10.1